connector: oauth2 do not fetch issuer configuration before checking claim
Note: it is still possible to call this code without a iss claim, in
that case the configuration is fetched and can be used to make
unsolicited calls to well-known configuration endpoints.  The
oauth2/bearer-token interceptor does require iss to be set so the
gateway will not introduce such loophole.